requirements.txt
Django>=3.1
django-cors-headers>=3.2.0
django_extensions
django-filter>=2.2.0
django-ipware>=2.1.0
djangorestframework>=3.9.0
django-model-utils>=4.1.1
docxtpl>=0.16.0
lxml>=4.2.5
openpyxl>=3.0.9
parse>=1.19.0
psycopg2>=2.7.5
pytest>=5.3.1
pytest-django>=3.7.0
reportlab>=3.6.9
sqlparams>=5.0.0
xlwt>=1.3.0
wget>=3.2
PyPDF2>=1.27.3
